+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fix overflow issue with `RelatedPage` in IE11. [#238](https://github.com/mapbox/dr-ui/pull/238)
 * Create Vimeo modal option (`vimeoId` and `vimeoThumbnail`) for `RelatedPage`. [#236](https://github.com/mapbox/dr-ui/pull/236)
+* Add playground theme from `RelatedPage`. [#234](https://github.com/mapbox/dr-ui/pull/234)
 
 ## 0.25.5